factor out gen selection slider
