+++ .eslintrc.json
@@ -19,7 +19,6 @@
     ],
     "rules": {
         "no-console": "off",
-        "react/jsx-props-no-spreading": "off",
-        "react/forbid-prop-types": "off" // TODO: remove need for this
+        "react/jsx-props-no-spreading": "off"
     }
 }
